Better actions, change the erase of selections
