+++ src/main/scala/io/atal/butterfly/Action/Action.scala
@@ -1,0 +1,7 @@
+package io.atal.butterfly
+
+/** Define an Action
+  */
+trait Action {
+  def execute(editor: Editor): Unit
+}
