+++ src/main/scala/io/atal/butterfly/Action/Copy.scala
@@ -1,0 +1,9 @@
+package io.atal.butterfly
+
+/** Implement the action Copy
+  * Copy the text of the selection
+  * If several selections, concat all the selection's texts
+  */
+class Copy() extends Action {
+  def execute(editor: Editor): Unit = Unit
+}
